pyproject.toml
[tool.poetry]
name = "superbenchmark"
version = "0.1.0"
description = ""
authors = ["Viktor <vktr.buyukli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
pydantic-settings = "^2.9.1"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
